+++ example/cloudflare_ai_example.dart
@@ -3,8 +3,8 @@ import 'dart:typed_data';
 import 'package:cloudflare_ai/cloudflare_ai.dart';
 
 void main() async {
-  String accountId = "YOUR_ACCOUNT_ID";
-  String apiKey = "YOUR_API_KEY";
+  String accountId = "";
+  String apiKey = "";
 
   // Text Generation
   // Initialize a TextGenerationModel
@@ -18,10 +18,10 @@ void main() async {
   TextGenerationResponse textGenRes = await textGenModel
       .generateText("Write a story about a robot, living on the moon");
 
-  if (textGenRes.success) {
-    print(textGenRes.result.response);
+  if (textGenRes.success && textGenRes.result != null) {
+    print(textGenRes.result?.response);
   } else {
-    print(textGenRes.errors);
+    print(textGenRes..errors.map((e) => e.toJson()).toList());
   }
 
   // Text Summarization
@@ -40,10 +40,10 @@ void main() async {
     maxLength: 1024, // 1024 is default, hence this parameter is optional
   );
 
-  if (textSummarizationRes.success) {
-    print(textSummarizationRes.result.response);
+  if (textSummarizationRes.success && textSummarizationRes.result != null) {
+    print(textSummarizationRes.result?.response);
   } else {
-    print(textSummarizationRes.errors);
+    print(textSummarizationRes..errors.map((e) => e.toJson()).toList());
   }
 
   // Text to Image
@@ -70,19 +70,42 @@ void main() async {
           .DISTILBERT_SST_2_INT8 // DISTILBERT_SST_2_INT8 is default, hence this parameter is optional
       );
 
-  // Initialize TextClassificationResponse
+  // Classify Text
   TextClassificationResponse textClassificationResponse =
       await textClassificationModel.classifyText(
     "Test Prompt",
   );
 
-  // Printing the output
-  if (textClassificationResponse.success) {
+  if (textClassificationResponse.success &&
+      textClassificationResponse.result != null) {
     print(
-        'Positive Confidence level: ${textClassificationResponse.result.positive}');
+        'Positive Confidence level: ${textClassificationResponse.result?.positive}');
     print(
-        'Negative Confidence level: ${textClassificationResponse.result.negative}');
+        'Negative Confidence level: ${textClassificationResponse.result?.negative}');
   } else {
-    print(textSummarizationRes.errors);
+    print(textSummarizationRes..errors.map((e) => e.toJson()).toList());
+  }
+
+  // Language Translation
+  // Initialize a LanguageTranslationModel
+  LanguageTranslationModel languageTranslationModel = LanguageTranslationModel(
+    accountId: accountId,
+    apiKey: apiKey,
+    model: LanguageTranslationModels
+        .M2M100_1_2B, // M2M100_1_2B is default, hence this parameter is optional
+  );
+
+  // Translate Text
+  LanguageTranslationResponse languageTranslationRes =
+      await languageTranslationModel.translate(
+    "Hello",
+    Languages.English,
+    Languages.Hindi,
+  );
+
+  if (languageTranslationRes.success && languageTranslationRes.result != null) {
+    print(languageTranslationRes.result?.response);
+  } else {
+    print(languageTranslationRes..errors.map((e) => e.toJson()).toList());
   }
 }