+++ lib/src/cloudflare_ai_base.dart
@@ -2,3 +2,4 @@ export 'text_generation/text_generation.dart';
 export 'text_summarization/text_summarization.dart';
 export 'text_to_image/text_to_image.dart';
 export 'text_classification/text_classification.dart';
+export 'language_translation/language_translation.dart';