+++ lib/src/text_generation/text_generation.dart
@@ -19,6 +19,12 @@ class TextGenerationModel {
     required this.model,
   }) {
     baseUrl = "https://api.cloudflare.com/client/v4/accounts/$accountId/ai/run";
+    if (accountId.trim() == "") {
+      throw Exception("Account ID cannot be empty");
+    }
+    if (apiKey.trim() == "") {
+      throw Exception("API Key cannot be empty");
+    }
   }
 
   // Asynchronous function which returns generated text through the TextGenerationResponse object